+++ debian-control/Cargo.toml
@@ -22,19 +22,20 @@ chrono = { version = ">=0.4, <0.5", optional = true }
 serde = { version = "1", optional = true }
 
 [features]
-default = ["chrono", "lossless"]
+default = ["chrono", "edit"]
 python-debian = ["dep:pyo3", "deb822-edit/python-debian"]
 chrono = ["dep:chrono"]
 serde = ["dep:serde"]
-lossless = ["dep:deb822-edit", "dep:rowan"]
+edit = ["dep:deb822-edit", "dep:rowan"]
+lossless = ["edit"]  # Deprecated: use 'edit' feature instead
 
 [[example]]
 name = "create-file"
-required-features = ["lossless"]
+required-features = ["edit"]
 
 [[example]]
 name = "incremental_parse"
-required-features = ["lossless"]
+required-features = ["edit"]
 
 [dev-dependencies]
 serde_json = "1.0.148"
