+++ debian-control/examples/incremental_parse.rs
@@ -1,5 +1,5 @@
 use deb822_edit::TextRange;
-use debian_control::lossless::Control;
+use debian_control::edit::Control;
 
 fn main() {
     // Example control file content
