+++ debian-control/src/edit/mod.rs
@@ -1,13 +1,13 @@
-//! Lossless parser for various Debian control files
+//! Editable parser for various Debian control files
 //!
 //! This library provides a parser for various Debian control files, such as `control`, `changes`,
-//! and apt `Release`, `Packages`, and `Sources` files. The parser is lossless, meaning that it
-//! preserves all formatting as well as any possible errors in the files.
+//! and apt `Release`, `Packages`, and `Sources` files. The parser preserves all formatting,
+//! allowing for precise edits while maintaining the original structure and any possible errors in the files.
 //!
 //! # Example with positioned errors
 //!
 //! ```
-//! use debian_control::lossless::{Control, PositionedParseError};
+//! use debian_control::edit::{Control, PositionedParseError};
 //!
 //! let input = "Invalid: field\nBroken field without colon";
 //! let parsed = Control::parse(input);
