+++ debian-control/src/edit/relations.rs
@@ -2,7 +2,7 @@
 //!
 //! # Example
 //! ```
-//! use debian_control::lossless::relations::{Relations, Relation};
+//! use debian_control::edit::relations::{Relations, Relation};
 //! use debian_control::relations::VersionConstraint;
 //!
 //! let mut relations: Relations = r"python3-dulwich (>= 0.19.0), python3-requests, python3-urllib3 (<< 1.26.0)".parse().unwrap();
@@ -1067,7 +1067,7 @@ impl Relations {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Relations;
+    /// use debian_control::edit::relations::Relations;
     ///
     /// let mut relations: Relations = "debhelper (>= 9)".parse().unwrap();
     /// relations.ensure_minimum_version("debhelper", &"12".parse().unwrap());
@@ -1166,7 +1166,7 @@ impl Relations {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Relations;
+    /// use debian_control::edit::relations::Relations;
     ///
     /// let mut relations: Relations = "debhelper (>= 9)".parse().unwrap();
     /// relations.ensure_exact_version("debhelper", &"12".parse().unwrap());
@@ -1230,7 +1230,7 @@ impl Relations {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Relations;
+    /// use debian_control::edit::relations::Relations;
     ///
     /// let mut relations: Relations = "python3".parse().unwrap();
     /// relations.ensure_some_version("debhelper");
@@ -1272,7 +1272,7 @@ impl Relations {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::{Relations, Entry};
+    /// use debian_control::edit::relations::{Relations, Entry};
     ///
     /// let mut relations: Relations = "python3".parse().unwrap();
     /// let new_entry: Entry = "debhelper (>= 12)".parse().unwrap();
@@ -1335,7 +1335,7 @@ impl Relations {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Relations;
+    /// use debian_control::edit::relations::Relations;
     ///
     /// let mut relations: Relations = "python3".parse().unwrap();
     /// relations.ensure_substvar("${misc:Depends}").unwrap();
@@ -1399,7 +1399,7 @@ impl Relations {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Relations;
+    /// use debian_control::edit::relations::Relations;
     ///
     /// let (mut relations, _) = Relations::parse_relaxed("python3, ${misc:Depends}", true);
     /// relations.drop_substvar("${misc:Depends}");
@@ -1472,7 +1472,7 @@ impl Relations {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Relations;
+    /// use debian_control::edit::relations::Relations;
     ///
     /// let mut relations: Relations = "python3, debhelper, rustc".parse().unwrap();
     /// relations.filter_entries(|entry| {
@@ -1503,7 +1503,7 @@ impl Relations {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::{Relations, WrapAndSortOrder};
+    /// use debian_control::edit::relations::{Relations, WrapAndSortOrder};
     ///
     /// let relations: Relations = "debhelper, python3, rustc".parse().unwrap();
     /// assert!(relations.is_sorted(&WrapAndSortOrder));
@@ -1627,7 +1627,7 @@ impl Relations {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Relations;
+    /// use debian_control::edit::relations::Relations;
     ///
     /// let mut relations: Relations = "python3, debhelper, rustc".parse().unwrap();
     /// assert!(relations.drop_dependency("debhelper"));
@@ -1672,7 +1672,7 @@ impl Relations {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::{Relations, Relation, Entry};
+    /// use debian_control::edit::relations::{Relations, Relation, Entry};
     ///
     /// let mut relations: Relations = "python3, rustc".parse().unwrap();
     /// let entry = Entry::from(Relation::simple("debhelper"));
@@ -1700,7 +1700,7 @@ impl Relations {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Relations;
+    /// use debian_control::edit::relations::Relations;
     ///
     /// let relations: Relations = "python3, debhelper (>= 12), rustc".parse().unwrap();
     /// let (idx, entry) = relations.get_relation("debhelper").unwrap();
@@ -1733,7 +1733,7 @@ impl Relations {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Relations;
+    /// use debian_control::edit::relations::Relations;
     ///
     /// let relations: Relations = "python3 | python3-minimal, python3-dev".parse().unwrap();
     /// let entries: Vec<_> = relations.iter_relations_for("python3").collect();
@@ -1757,7 +1757,7 @@ impl Relations {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Relations;
+    /// use debian_control::edit::relations::Relations;
     ///
     /// let relations: Relations = "python3, debhelper, rustc".parse().unwrap();
     /// assert!(relations.has_relation("debhelper"));
@@ -1934,7 +1934,7 @@ impl Entry {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::{Relation,Entry};
+    /// use debian_control::edit::relations::{Relation,Entry};
     /// let mut entry: Entry = r"python3-dulwich (>= 0.19.0) | python3-requests".parse().unwrap();
     /// entry.remove_relation(1);
     /// assert_eq!(entry.to_string(), "python3-dulwich (>= 0.19.0)");
@@ -1952,7 +1952,7 @@ impl Entry {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::{Relation,Entry};
+    /// use debian_control::edit::relations::{Relation,Entry};
     /// let entry = Entry::from(vec!["samba (>= 2.0)".parse::<Relation>().unwrap()]);
     /// assert!(entry.satisfied_by(|name: &str| -> Option<debversion::Version> {
     ///    match name {
@@ -1988,7 +1988,7 @@ impl Entry {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::{Relations,Entry};
+    /// use debian_control::edit::relations::{Relations,Entry};
     /// let mut relations: Relations = r"python3-dulwich (>= 0.19.0), python3-urllib3 (<< 1.26.0)".parse().unwrap();
     /// let mut entry = relations.get_entry(0).unwrap();
     /// entry.remove();
@@ -2052,7 +2052,7 @@ impl Entry {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::{Relation,Entry};
+    /// use debian_control::edit::relations::{Relation,Entry};
     /// let mut entry: Entry = "samba (>= 2.0)".parse().unwrap();
     /// entry.push("python3-requests".parse().unwrap());
     /// assert_eq!(entry.to_string(), "samba (>= 2.0) | python3-requests");
@@ -2132,7 +2132,7 @@ impl Entry {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Entry;
+    /// use debian_control::edit::relations::Entry;
     ///
     /// let inner: Entry = "pkg (>= 1.0)".parse().unwrap();
     /// let outer: Entry = "pkg (>= 1.5) | libc6".parse().unwrap();
@@ -2222,7 +2222,7 @@ impl Relation {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::{Relation};
+    /// use debian_control::edit::relations::{Relation};
     /// use debian_control::relations::VersionConstraint;
     /// let relation = Relation::new("samba", Some((VersionConstraint::GreaterThanEqual, "2.0".parse().unwrap())));
     /// assert_eq!(relation.to_string(), "samba (>= 2.0)");
@@ -2266,7 +2266,7 @@ impl Relation {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Relation;
+    /// use debian_control::edit::relations::Relation;
     /// let relation = "  samba  (  >= 2.0) ".parse::<Relation>().unwrap();
     /// assert_eq!(relation.wrap_and_sort().to_string(), "samba (>= 2.0)");
     /// ```
@@ -2344,7 +2344,7 @@ impl Relation {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Relation;
+    /// use debian_control::edit::relations::Relation;
     /// let relation = Relation::simple("samba");
     /// assert_eq!(relation.to_string(), "samba");
     /// ```
@@ -2356,7 +2356,7 @@ impl Relation {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::{Relation};
+    /// use debian_control::edit::relations::{Relation};
     /// use debian_control::relations::VersionConstraint;
     /// let mut relation = Relation::new("samba", Some((VersionConstraint::GreaterThanEqual, "2.0".parse().unwrap())));
     /// relation.drop_constraint();
@@ -2387,7 +2387,7 @@ impl Relation {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Relation;
+    /// use debian_control::edit::relations::Relation;
     /// let relation = Relation::simple("samba");
     /// assert_eq!(relation.try_name(), Some("samba".to_string()));
     /// ```
@@ -2403,7 +2403,7 @@ impl Relation {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Relation;
+    /// use debian_control::edit::relations::Relation;
     /// let relation: Relation = "samba (>= 4.0)".parse().unwrap();
     /// let range = relation.name_range().unwrap();
     /// assert_eq!(&relation.to_string()[range], "samba");
@@ -2412,7 +2412,7 @@ impl Relation {
         self.name_token().map(|token| token.text_range())
     }
 
-    fn name_token(&self) -> Option<crate::lossless::relations::SyntaxToken> {
+    fn name_token(&self) -> Option<crate::edit::relations::SyntaxToken> {
         self.0.children_with_tokens().find_map(|it| match it {
             SyntaxElement::Token(token) if token.kind() == IDENT => Some(token),
             _ => None,
@@ -2427,7 +2427,7 @@ impl Relation {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Relation;
+    /// use debian_control::edit::relations::Relation;
     /// let relation = Relation::simple("samba");
     /// assert_eq!(relation.name(), "samba");
     /// ```
@@ -2443,7 +2443,7 @@ impl Relation {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Relation;
+    /// use debian_control::edit::relations::Relation;
     /// let relation: Relation = "samba:any".parse().unwrap();
     /// assert_eq!(relation.archqual(), Some("any".to_string()));
     /// ```
@@ -2464,7 +2464,7 @@ impl Relation {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Relation;
+    /// use debian_control::edit::relations::Relation;
     /// let mut relation = Relation::simple("samba");
     /// relation.set_archqual("any");
     /// assert_eq!(relation.to_string(), "samba:any");
@@ -2529,7 +2529,7 @@ impl Relation {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::{Relation};
+    /// use debian_control::edit::relations::{Relation};
     /// use debian_control::relations::VersionConstraint;
     /// let mut relation = Relation::simple("samba");
     /// relation.set_version(Some((VersionConstraint::GreaterThanEqual, "2.0".parse().unwrap())));
@@ -2619,7 +2619,7 @@ impl Relation {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Relation;
+    /// use debian_control::edit::relations::Relation;
     /// let relation: Relation = "samba [amd64]".parse().unwrap();
     /// assert_eq!(relation.architectures().unwrap().collect::<Vec<_>>(), vec!["amd64".to_string()]);
     /// ```
@@ -2640,7 +2640,7 @@ impl Relation {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::{Relation};
+    /// use debian_control::edit::relations::{Relation};
     /// use debian_control::relations::{BuildProfile};
     /// let relation: Relation = "samba <!nocheck>".parse().unwrap();
     /// assert_eq!(relation.profiles().collect::<Vec<_>>(), vec![vec![BuildProfile::Disabled("nocheck".to_string())]]);
@@ -2682,7 +2682,7 @@ impl Relation {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Relation;
+    /// use debian_control::edit::relations::Relation;
     /// let relation: Relation = "samba <!nocheck>".parse().unwrap();
     /// let ranges: Vec<_> = relation.profile_ranges().collect();
     /// assert_eq!(ranges.len(), 1);
@@ -2707,7 +2707,7 @@ impl Relation {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::{Relation,Entry};
+    /// use debian_control::edit::relations::{Relation,Entry};
     /// let mut entry: Entry = r"python3-dulwich (>= 0.19.0) | python3-urllib3 (<< 1.26.0)".parse().unwrap();
     /// let mut relation = entry.get_relation(0).unwrap();
     /// relation.remove();
@@ -2777,7 +2777,7 @@ impl Relation {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Relation;
+    /// use debian_control::edit::relations::Relation;
     /// let mut relation = Relation::simple("samba");
     /// relation.set_architectures(vec!["amd64", "i386"].into_iter());
     /// assert_eq!(relation.to_string(), "samba [amd64 i386]");
@@ -2835,7 +2835,7 @@ impl Relation {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Relation;
+    /// use debian_control::edit::relations::Relation;
     /// use debian_control::relations::BuildProfile;
     /// let mut relation = Relation::simple("samba");
     /// relation.add_profile(&[BuildProfile::Disabled("nocheck".to_string())]);
@@ -2914,7 +2914,7 @@ impl Relation {
     ///
     /// # Example
     /// ```
-    /// use debian_control::lossless::relations::Relation;
+    /// use debian_control::edit::relations::Relation;
     /// use debian_control::relations::VersionConstraint;
     ///
     /// let inner = Relation::new("pkg", Some((VersionConstraint::GreaterThanEqual, "1.0".parse().unwrap())));
@@ -2983,7 +2983,7 @@ impl Relation {
 ///
 /// # Example
 /// ```
-/// use debian_control::lossless::relations::{Relation};
+/// use debian_control::edit::relations::{Relation};
 /// use debian_control::relations::VersionConstraint;
 /// let relation = Relation::build("samba")
 ///    .version_constraint(VersionConstraint::GreaterThanEqual, "2.0".parse().unwrap())
@@ -4001,7 +4001,7 @@ mod tests {
     #[test]
     fn test_ensure_minimum_version_preserves_newline_in_control() {
         // Test the full scenario from the bug report
-        use crate::lossless::Control;
+        use crate::edit::Control;
         use std::str::FromStr;
 
         let input = r#"Source: f2fs-tools
