+++ debian-control/src/lib.rs
@@ -33,21 +33,30 @@
 //! "#);
 //! ```
 //!
-//! See the ``lossless`` module for a parser that preserves all comments and formatting, and
+//! See the ``edit`` module for a parser that preserves all comments and formatting, and
 //! as well as allowing inline errors.
 pub mod lossy;
-#[cfg(feature = "lossless")]
-pub use lossless::control::{Binary, Control, Source};
+#[cfg(feature = "edit")]
+pub use edit::control::{Binary, Control, Source};
 pub mod fields;
 pub use fields::*;
+#[cfg(feature = "edit")]
+pub mod edit;
+#[cfg(feature = "edit")]
+pub use edit::apt;
+#[cfg(feature = "edit")]
+pub use edit::changes;
+#[cfg(feature = "edit")]
+pub use edit::control;
+
 #[cfg(feature = "lossless")]
-pub mod lossless;
-#[cfg(feature = "lossless")]
-pub use lossless::apt;
-#[cfg(feature = "lossless")]
-pub use lossless::changes;
-#[cfg(feature = "lossless")]
-pub use lossless::control;
+#[deprecated(since = "0.1.49", note = "Use `edit` module instead")]
+pub mod lossless {
+    //! Deprecated: Use the `edit` module instead.
+    //!
+    //! This module is a compatibility alias for the `edit` module and will be removed in a future version.
+    pub use crate::edit::*;
+}
 #[cfg(feature = "lossless")]
 pub mod pgp;
 pub mod relations;
