+++ debian-control/src/relations.rs
@@ -188,7 +188,7 @@ pub enum SyntaxKind {
 }
 
 /// Convert our `SyntaxKind` into the rowan `SyntaxKind`.
-#[cfg(feature = "lossless")]
+#[cfg(feature = "edit")]
 impl From<SyntaxKind> for rowan::SyntaxKind {
     fn from(kind: SyntaxKind) -> Self {
         Self(kind as u16)
