+++ debian-copyright/Cargo.toml
@@ -17,8 +17,9 @@ deb822-edit = { version = ">=0.5.16, <0.6", path = "../deb822-edit", optional = 
 deb822-fast = { version = ">=0.1, <0.3", path = "../deb822-fast", features = ["derive"] }
 
 [features]
-default = ["lossless"]
-lossless = ["dep:deb822-edit"]
+default = ["edit"]
+edit = ["dep:deb822-edit"]
+lossless = ["edit"]  # Deprecated: use 'edit' feature instead
 
 [badges]
 maintenance = { status = "actively-maintained" }
